old/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rss2article"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langgraph==1.0.5",
  "langchain-community==0.4.1",
  "feedparser==6.0.12",
  "listparser==0.20",
  "newspaper3k==0.2.8",
  "lxml-html-clean==0.4.3",
  "langchain-openai==1.1.6",
  "python-dotenv==1.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
buzzter_agent = ["py.typed"]

[tool.mypy]
mypy_path = "src"
packages = ["buzzter_agent"]
